Reduce code duplication in MPI partitioning
